Migrated devmodules and animations
