+++ apps/webclient/src/app/app.component.ts
@@ -1,9 +1,11 @@
 import { ChangeDetectionStrategy, Component } from '@angular/core';
+import { RouterOutlet } from '@angular/router';
 
 @Component({
   standalone: true,
   selector: 'webclient-root',
-  template: '<h1>TEST</h1>',
+  imports: [RouterOutlet],
+  template: '<router-outlet></router-outlet>',
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class AppComponent {}
