--- apps/webclient/src/app/app.config.ts
@@ -1,5 +1,0 @@
-import { ApplicationConfig } from '@angular/core';
-
-export const appConfig: ApplicationConfig = {
-  providers: [],
-};
