+++ apps/webclient/src/main.ts
@@ -1,5 +1,5 @@
 import { bootstrapApplication } from '@angular/platform-browser';
-import { appConfig } from './app/app.config';
 import { AppComponent } from './app/app.component';
+import { DesktopShellConfig } from '@towech-finance/desktop/shell/feature';
 
-bootstrapApplication(AppComponent, appConfig).catch(err => console.error(err));
+bootstrapApplication(AppComponent, DesktopShellConfig).catch(err => console.error(err));
