+++ libs/desktop/shell/feature/.eslintrc.json
