+++ libs/desktop/shell/feature/environments/environment.prod.ts
