+++ libs/desktop/shell/feature/environments/environment.spec.ts
