+++ libs/desktop/shell/feature/environments/environment.ts
