+++ libs/desktop/shell/feature/jest.config.ts
