+++ libs/desktop/shell/feature/project.json
