+++ libs/desktop/shell/feature/src/lib/desktop-shell.module.ts
@@ -1,0 +1,14 @@
+// TODO: Eventually remove this
+@NgModule({
+  imports: [
+    // Base
+    CommonModule,
+    // Http
+    HttpClientModule,
+    // NGRX
+    DesktopShellDataAccessUserStateModule,
+    // Guards
+    DesktopShellUtilsGuardsModule,
+  ],
+})
+export class DesktopShellModule {}
