+++ libs/desktop/shell/feature/tsconfig.json
