+++ libs/desktop/shell/feature/tsconfig.lib.json
