--- libs/desktop_old/shell/feature/src/index.ts
@@ -1,1 +1,0 @@
-export * from './lib/desktop-shell.module';
