--- libs/desktop_old/shell/utils/environments/.eslintrc.json
@@ -1,33 +1,0 @@
-{
-  "extends": ["../../../../../.eslintrc.json"],
-  "ignorePatterns": ["!**/*"],
-  "overrides": [
-    {
-      "files": ["*.ts"],
-      "rules": {
-        "@angular-eslint/directive-selector": [
-          "error",
-          {
-            "type": "attribute",
-            "prefix": "towechFinance",
-            "style": "camelCase"
-          }
-        ],
-        "@angular-eslint/component-selector": [
-          "error",
-          {
-            "type": "element",
-            "prefix": "towech-finance",
-            "style": "kebab-case"
-          }
-        ]
-      },
-      "extends": ["plugin:@nx/angular", "plugin:@angular-eslint/template/process-inline-templates"]
-    },
-    {
-      "files": ["*.html"],
-      "extends": ["plugin:@nx/angular-template"],
-      "rules": {}
-    }
-  ]
-}
