--- libs/desktop_old/shell/utils/environments/README.md
@@ -1,7 +1,0 @@
-# desktop-shell-utils-environments
-
-This library was generated with [Nx](https://nx.dev).
-
-## Running unit tests
-
-Run `nx test desktop-shell-utils-environments` to execute the unit tests.
