--- libs/desktop_old/shell/utils/environments/src/index.ts
@@ -1,2 +1,0 @@
-export * from './lib/desktop-shell-utils-environments.module';
-export * from './lib/environments/environment';
