--- libs/desktop_old/shell/utils/environments/src/test-setup.ts
@@ -1,1 +1,0 @@
-import 'jest-preset-angular/setup-jest';
